新增加入空行的功能，修正資料形態轉換的問題

+++ txtFile/txtFile.cpp
@@ -101,7 +101,7 @@ int txtFile::print() {
 }
 
 unsigned long txtFile::getNLine() {
-    return TXTContent.size();
+    return (unsigned long)TXTContent.size();
 }
 
 std::string txtFile::getLineContent(unsigned long NoLine) {
